For Windows, download terminal-config into initial folder

+++ windows.bat
@@ -1,9 +1,9 @@
 if not exist "C:\Users\%USERNAME%\tmp" mkdir C:\Users\%USERNAME%\tmp
 
-if exist "C:\Users\%USERNAME%\tmp\terminal-config" rmdir /s /q C:\Users\%USERNAME%\tmp\terminal-config
-git clone --depth=1 git@github.com:johnnyodonnell/terminal-config.git C:\Users\%USERNAME%\tmp\terminal-config
+if exist "C:\Users\%USERNAME%\tmp\terminal-config-initial" rmdir /s /q C:\Users\%USERNAME%\tmp\terminal-config-initial
+git clone --depth=1 git@github.com:johnnyodonnell/terminal-config.git C:\Users\%USERNAME%\tmp\terminal-config-initial
 
-"C:\Users\%USERNAME%\tmp\terminal-config\installers\windows.bat"
+"C:\Users\%USERNAME%\tmp\terminal-config-initial\installers\windows.bat"
 
-rmdir /s /q C:\Users\%USERNAME%\tmp\terminal-config
+rmdir /s /q C:\Users\%USERNAME%\tmp\terminal-config-initial
